requirements.txt
pandas>=2.0.0
mysql-connector-python>=8.4.0
python-dotenv>=1.0.0
